requirements.txt
pyqt6
opencv-python
pygame
websockets
numpy
Pillow
av>=10.0.0             # PyAV for GPU-accelerated video encoding (NVENC)

# Real device control dependencies
bleak>=0.21.0          # Bluetooth Low Energy support
cffi>=1.15.0           # C Foreign Function Interface for Rust bindings

# Testing dependencies
pytest
pytest-asyncio
pytest-timeout
pytest-qt

PyOpenGL>=3.1.6
PyOpenGL_accelerate>=3.1.6

